+++ src/js/tools/transformation/text/camelCaseConverter.js
@@ -1,0 +1,3 @@
+export default function camelCaseConverter(string, delimeter = "-") {
+    return string.replace(RegExp(`${delimeter}(.)`, "g"), (match, p1) => p1.toUpperCase())
+}