+++ src/js/ui/Animation/Classes/Animation.js
@@ -3,26 +3,35 @@ import FieldChecker from "../../../tools/validation/fieldChecker"
 import Linear from "../Library/Timing/linear"
 
 export default class Animation {
-    constructor({ duration, painter, timingFunc = Linear }) {
+    constructor({
+        duration, painter, init = () => { }, end = () => { }, timingFunc = Linear,
+    }) {
         new FieldsContainer([
             ["duration", "painter", "timingFunc"],
             {
                 duration: new FieldChecker({ type: "number", isint: "true" }),
                 painter: new FieldChecker({ type: "function" }),
                 timingFunc: new FieldChecker({ type: "function" }),
+                init: new FieldChecker({ type: "function" }),
+                end: new FieldChecker({ type: "function" }),
             },
-        ]).set({ duration, painter, timingFunc })
+        ]).set({
+            duration, painter, timingFunc, init, end,
+        })
 
         this.duration = duration
         this.painter = painter
         this.timingFunc = timingFunc
+        this.init = init
+        this.end = end
     }
 
     animate(element) {
-        return new Promise((resolve, reject) => {
+        return new Promise(async (resolve, reject) => {
+            const init = await this.init(element)
             const start = performance.now()
 
-            let animate = (time) => {
+            let animate = async (time) => {
                 let timeProgress = (time - start) / this.duration
                 if (timeProgress > 1) timeProgress = 1
 
@@ -30,13 +39,16 @@ export default class Animation {
                 this.painter.bind(element)(effectProgress)
 
                 if (timeProgress < 1) {
-                    requestAnimationFrame(animate)
-                } else resolve(element)
+                    requestAnimationFrame(animate, init)
+                } else {
+                    await this.end(element)
+                    resolve(element)
+                }
             }
 
             animate = animate.bind(this)
 
-            requestAnimationFrame(animate)
+            requestAnimationFrame(animate, init)
         })
     }
 
@@ -44,7 +56,7 @@ export default class Animation {
         return this.animate(...data)
     }
 
-    applyCallback({ data = [], callback = () => {} }) {
+    applyCallback({ data = [], callback = () => { } }) {
         this.animate(...data)
             .then(e => callback(e))
     }