+++ src/js/ui/DOM/Includes/Modificators/Sizes.js
@@ -1,0 +1,8 @@
+import DOMController from "../../Helpers/domController"
+
+DOMController.registerModificator({
+    name: "sizes",
+    get() {
+        return this.elementParse.native.getBoundingClientRect()
+    },
+})